// apps/wbs-planner/src/generated/models/risk-model.ts
import type { Resource } from './resource-model';
import type { Project } from './project-model';
import type { WBSItem } from './wbs-item-model';

export const RiskImpactKeyToLabel = {
  'Low': 'Low',
  'Medium': 'Medium',
  'High': 'High'
} as const;
export type RiskImpactKey = keyof typeof RiskImpactKeyToLabel;

export const RiskProbabilityKeyToLabel = {
  'Low': 'Low',
  'Medium': 'Medium',
  'High': 'High'
} as const;
export type RiskProbabilityKey = keyof typeof RiskProbabilityKeyToLabel;

export const RiskStatusKeyToLabel = {
  'Open': 'Open',
  'Mitigating': 'Mitigating',
  'Closed': 'Closed',
  'Accepted': 'Accepted',
  'Escalated': 'Escalated'
} as const;
export type RiskStatusKey = keyof typeof RiskStatusKeyToLabel;

export interface Risk {
  /**
   * @displayName Risk
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Risk Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  riskName: string;
  /**
   * @displayName Closed Date
   */
  closedDate?: string;
  /**
   * @displayName Created Date
   * @validationRule Required for create/update operations
   */
  createdDate: string;
  /**
   * @displayName Due Date
   * @validationRule Required for create/update operations
   */
  dueDate: string;
  /**
   * @displayName Impact
   * @validationRule Required for create/update operations
   */
  impactKey: RiskImpactKey;
  /**
   * @displayName Mitigation Plan
   */
  mitigationPlan?: string;
  /**
   * @displayName Owner
   * @validationRule Required for create/update operations
   */
  owner: Pick<Resource, 'id' | 'resourceName'>;
  /**
   * @displayName Probability
   * @validationRule Required for create/update operations
   */
  probabilityKey: RiskProbabilityKey;
  /**
   * @displayName Project
   * @validationRule Required for create/update operations
   */
  project: Pick<Project, 'id' | 'projectName'>;
  /**
   * @displayName Risk Score
   * @validationRule Required for create/update operations
   */
  riskScore: number;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: RiskStatusKey;
  /**
   * @displayName WBS Item
   */
  wBSItem?: Pick<WBSItem, 'id' | 'wBSItemName'>;
}

export const _Risk = 'Risk' as const;